+++ src/Exceptions/TransitionNotAllowedException.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the TransitionNotAllowedException class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-03-13
+ *
+ */
+
+namespace Vanilo\Workflow\Exceptions;
+
+use RuntimeException;
+
+class TransitionNotAllowedException extends RuntimeException
+{
+}