+++ Changelog.md
@@ -4,3 +4,4 @@
 ##### 2023-XX-YY
 
 - Initial release
+- Supports native PHP 8.1 BackedEnum and Konekt Enum (v3 & v4) classes